[fix] 풍향 계산기능 구현 및 오늘날짜 표시

+++ WetherApp/Model/CurrentWeatherResult.swift
@@ -33,7 +33,7 @@ struct WeatherMain: Codable {
 
 struct Wind: Codable {
     let speed: Double // 풍속 (m/s)
-    let deg: Int // 풍향 (도)
+    let deg: Double // 풍향 (도)
     let gust: Double? // 돌풍 속도 (옵셔널)
 }
